removed error in runUnsat.sh

+++ benchmarks/BV-benchamrks/mot-example/libt.c
@@ -1,4 +1,4 @@
 
 #include "libt.h"
 
-uint32_t _mul_cb(uint32_t a, uint32_t b) { return a * b; }
+uint8_t _mul_cb(uint8_t a, uint8_t b) { return a * b; }
